src/pr_git/requirements.txt
fastapi
uvicorn[standard]
boto3
requests
python-dateutil
